% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_props.R
\name{add_title_pos}
\alias{add_title_pos}
\title{add_title_pos}
\usage{
add_title_pos(vis, ..., title = "Plot Title", align = "left", angle = 0,
  dx = 0, dy = 0)
}
\arguments{
\item{vis, }{a ggvis plot}

\item{..., }{any additional arguments}

\item{title}{for the plot}

\item{align}{position of the title c('left','right')}

\item{angle}{of the labels in x axis}

\item{dx, }{relative x position of the labels in the x axis}

\item{dy, }{relative y position of the labels in the x axis}
}
\value{
deseq2 results
}
\description{
Adds a title with extra axis to ggvis plot and sets the positions
}
\examples{
    require(ggvis)
    mtcars \%>\%
    ggvis(x=~cyl, y=~wt, fill=~mpg) \%>\%
    group_by(mpg) \%>\%
    layer_bars() \%>\%
    add_title_pos(title = "title", angle=310, dy=0, dx=0) \%>\%
    set_options(width = 400, height = 350)

}
