% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{addID}
\alias{addID}
\title{addID}
\usage{
addID(data = NULL)
}
\arguments{
\item{data, }{loaded dataset}
}
\value{
data
}
\description{
Adds an id to the data frame being used.
}
\examples{
    x <- addID()

}
