% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all2all.R
\name{all2all}
\alias{all2all}
\title{all2all}
\usage{
all2all(data, cex = 2)
}
\arguments{
\item{data, }{data that have the sample names in the header.}

\item{cex}{text size}
}
\value{
all2all scatter plots
}
\description{
Prepares all2all scatter plots for given datasets.
}
\examples{
    plot<-all2all(mtcars)

}
