% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{applyFilters}
\alias{applyFilters}
\title{applyFilters}
\usage{
applyFilters(filt_data = NULL, cols = NULL, input = NULL)
}
\arguments{
\item{filt_data, }{loaded dataset}

\item{cols, }{selected samples}

\item{input, }{input parameters}
}
\value{
data
}
\description{
applyFilters
}
\examples{
    x <- applyFilters()

}

