% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{axis_vars}
\alias{axis_vars}
\title{Global variable axis_vars}
\format{An object of class \code{character} of length 2.}
\usage{
axis_vars
}
\value{
axis_vars
}
\description{
Global variable axis_var for axis variables of x and y
}
\examples{
    axis_vars <- c(x, y)

}
\keyword{datasets}

