% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batcheffect.R
\name{batchEffectUI}
\alias{batchEffectUI}
\title{batchEffectUI
Creates a panel to coorect batch effect}
\usage{
batchEffectUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
panel
}
\description{
batchEffectUI
Creates a panel to coorect batch effect
}
\examples{
    x <- batchEffectUI("batcheffect")

}
