% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark.R
\name{bookmarkServer}
\alias{bookmarkServer}
\title{bookmarkServer}
\usage{
bookmarkServer(input = NULL, output = NULL, session = NULL,
  loadingJSON = NUL)
}
\arguments{
\item{input, }{input}

\item{output, }{output}

\item{session, }{session}

\item{loadingJSON, }{loadingJSON}
}
\description{
bookmark Server functions
}
\examples{
    x <- bookmarkServer()

}
