% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxmain.R
\name{BoxMainPlotControlsUI}
\alias{BoxMainPlotControlsUI}
\title{BoxMainPlotControlsUI}
\usage{
BoxMainPlotControlsUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
returns the controls for left menu
}
\description{
Generates the controls in the left menu for a Box main plot
}
\note{
\code{BoxMainPlotControlsUI}
}
\examples{
    x <- BoxMainPlotControlsUI("box")
}
