% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{clusterData}
\alias{clusterData}
\title{clusterData}
\usage{
clusterData(dat)
}
\arguments{
\item{dat, }{the data to cluster}
}
\value{
clustered data
}
\description{
clusterData
}
\note{
\code{clusterData}
}
\examples{
    mycluster <- clusterData(mtcars)

}

