% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{clustFunParamsUI}
\alias{clustFunParamsUI}
\title{clustFunParamsUI}
\usage{
clustFunParamsUI()
}
\value{
cluster params
}
\description{
get cluster function parameter control
}
\note{
\code{clustFunParamsUI}
}
\examples{
    x <- clustFunParamsUI()
}
