% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{condSelectUI}
\alias{condSelectUI}
\title{condSelectUI
Creates a panel to select samples for each condition}
\usage{
condSelectUI()
}
\value{
panel
}
\description{
condSelectUI
Creates a panel to select samples for each condition
}
\examples{
    x <- condSelectUI()

}
