% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark.R
\name{copy2newDirectory}
\alias{copy2newDirectory}
\title{copy2newDirectory}
\usage{
copy2newDirectory(new_state_id = NULL, username = NULL, session = NULL)
}
\arguments{
\item{new_state_id, }{new state id}

\item{username, }{username}

\item{session, }{session}
}
\description{
To copy the bookmarked folder into a user named directory
}
\examples{
    x <- copy2newDirectory()

}
