% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{correctBatchEffect}
\alias{correctBatchEffect}
\title{Correct Batch Effect}
\usage{
correctBatchEffect(idata = NULL, input = NULL)
}
\arguments{
\item{idata, }{data}

\item{input, }{input values}
}
\value{
data
}
\description{
Batch effect correction
}
\examples{
    x<-correctBatchEffect ()
}
