% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{customColorsUI}
\alias{customColorsUI}
\title{customColorsUI}
\usage{
customColorsUI(id)
}
\arguments{
\item{id, }{namespace ID}
}
\value{
color range
}
\description{
get Custom Color controls
}
\note{
\code{getColRng}
}
\examples{
    x <- customColorsUI("heatmap")
}
