% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{dataLoadUI}
\alias{dataLoadUI}
\title{dataLoadUI}
\usage{
dataLoadUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
panel
}
\description{
Creates a panel to upload the data
}
\examples{
    x <- dataLoadUI("load")

}
