% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batcheffect.R
\name{debrowserbatcheffect}
\alias{debrowserbatcheffect}
\title{debrowserbatcheffect}
\usage{
debrowserbatcheffect(input, output, session, ldata = NULL)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{ldata, }{loaded data}
}
\value{
main plot

panel
}
\description{
Module to correct batch effect
}
\examples{
    x <- debrowserbatcheffect()

}
