% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{debrowserdeanalysis}
\alias{debrowserdeanalysis}
\title{debrowserdeanalysis}
\usage{
debrowserdeanalysis(input, output, session, data = NULL, columns = NULL,
  conds = NULL, params = NULL)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{data, }{a matrix that includes expression values}

\item{columns, }{columns}

\item{conds, }{conditions}

\item{params, }{de parameters}
}
\value{
DE panel
}
\description{
Module to perform and visualize DE results.
}
\examples{
    x <- debrowserdeanalysis()

}
