% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{debrowserpcaplot}
\alias{debrowserpcaplot}
\title{debrowserpcaplot}
\usage{
debrowserpcaplot(input = NULL, output = NULL, session = NULL,
  pcadata = NULL, metadata = NULL)
}
\arguments{
\item{input, }{input variables}

\item{output, }{output objects}

\item{session, }{session}

\item{pcadata, }{a matrix that includes expression values}

\item{metadata, }{metadata to color the plots}
}
\value{
main plot

panel
}
\description{
Module for a pca plot with its loadings 
as a mainplot in debrowser
}
\examples{
    x <- debrowserpcaplot()

}
