% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{deUI}
\alias{deUI}
\title{shinyUI to be able to run interactively}
\usage{
deUI()
}
\value{
the panel for main plots;
}
\description{
shinyUI to be able to run interactively
}
\note{
\code{deUI}
}
\examples{
    x<-deUI()

}

