% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{drawPCAExplained}
\alias{drawPCAExplained}
\title{drawPCAExplained}
\usage{
drawPCAExplained(explainedData = NULL)
}
\arguments{
\item{explainedData, }{selected data}
}
\value{
explained plot
}
\description{
Creates a more detailed plot using the PCA results from
the selected dataset.
}
\examples{
    x <- drawPCAExplained()

}

