% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{fileTypes}
\alias{fileTypes}
\title{fileTypes}
\usage{
fileTypes()
}
\value{
file types
}
\description{
Returns fileTypes that are going to be used in creating fileUpload UI
}
\note{
\code{fileTypes}
}
\examples{
    x <- fileTypes()

}
