% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{fileUploadBox}
\alias{fileUploadBox}
\title{fileUploadBox}
\usage{
fileUploadBox(id = NULL, inputId = NULL, label = NULL)
}
\arguments{
\item{id, }{namespace id}

\item{inputId, }{input file ID}

\item{label, }{label}
}
\value{
radio control
}
\description{
File upload module
}
\note{
\code{fileUploadBox}
}
\examples{
   
    x <- fileUploadBox("meta", "metadata", "Metadata")

}
