% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{generateTestData}
\alias{generateTestData}
\title{generateTestData}
\usage{
generateTestData(dat = NULL)
}
\arguments{
\item{dat, }{DESeq results will be generated for loaded data}
}
\value{
testData
}
\description{
This generates a test data that is suitable to main plots in debrowser
}
\examples{
    x <- generateTestData()

}
