% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark.R
\name{get_state_id}
\alias{get_state_id}
\title{get_state_id}
\usage{
get_state_id(prev_url = NULL)
}
\arguments{
\item{prev_url, }{prev url}
}
\description{
Helper to copy the bookmark to a user named directory
}
\examples{
    x <- get_state_id()

}
