% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_user_info}
\alias{get_user_info}
\title{Get the logged in user's email and other info}
\usage{
get_user_info(id = "me")
}
\arguments{
\item{id}{ID of the person to get the profile data for. 'me' to get current user.}
}
\value{
A People resource

https://developers.google.com/+/web/api/rest/latest/people#resource-representations
}
\description{
Get the logged in user's email and other info
}
\examples{

\dontrun{
options(googleAuthR.scopes.selected = 
   c("https://www.googleapis.com/auth/userinfo.email",
     "https://www.googleapis.com/auth/userinfo.profile"))
                                        
googleAuthR::gar_auth()

## default is user logged in
user <- get_user_info()
}

}
\seealso{
https://developers.google.com/+/web/api/rest/latest/people
}
