% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSection.R
\name{getAddPanel}
\alias{getAddPanel}
\title{getAddPanel, conditional panel for additional plots}
\usage{
getAddPanel()
}
\value{
the panel for additional plots;
}
\description{
getAddPanel, conditional panel for additional plots
}
\note{
\code{getAddSection}
}
\examples{
    x <- getAddPanel()

}

