% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSection.R
\name{getAddPlots}
\alias{getAddPlots}
\title{getAddPlots, for additional plots}
\usage{
getAddPlots(dataset, datasetname = "Up", addplot = "heatmap",
  metadata = NULL, clustering_method = "complete",
  distance_method = "cor", cex = 2)
}
\arguments{
\item{dataset, }{the dataset to use}

\item{datasetname, }{name of the dataset}

\item{addplot, }{type of plot to add}

\item{metadata, }{coupled samples and conditions}

\item{clustering_method, }{clustering method used}

\item{distance_method, }{distance method used}

\item{cex, }{font size}
}
\value{
the panel for additional plots;
}
\description{
getAddPlots, for additional plots
}
\note{
\code{getAddPlots}
}
\examples{
    x <- getAddPlots(mtcars)

}

