% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainpanel.R
\name{getAfterLoad}
\alias{getAfterLoad}
\title{getAfterLoad, after loading the data this text will be shown}
\usage{
getAfterLoad()
}
\value{
after load text
}
\description{
getAfterLoad, after loading the data this text will be shown
}
\note{
\code{getAfterLoad}
}
\examples{
 
    x <- getAfterLoad()

}

