% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all2all.R
\name{getAll2AllPlotUI}
\alias{getAll2AllPlotUI}
\title{getAll2AllPlotUI}
\usage{
getAll2AllPlotUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
the panel for all2all plots;
}
\description{
all2all plots UI.
}
\note{
\code{getAll2AllPlotUI}
}
\examples{
    x <- getAll2AllPlotUI("bar")

}
