% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barmain.R
\name{getBarMainPlot}
\alias{getBarMainPlot}
\title{getBarMainPlot}
\usage{
getBarMainPlot(data = NULL, conds = NULL, cols = NULL, key = NULL,
  title = "")
}
\arguments{
\item{data, }{count or normalized data}

\item{conds, }{conds}

\item{cols, }{cols}

\item{key, }{key}

\item{title, }{title}
}
\description{
Makes Density plots
}
\examples{
    getBarMainPlot()

}
