% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barmain.R
\name{getBarMainPlotUI}
\alias{getBarMainPlotUI}
\title{getBarMainPlotUI}
\usage{
getBarMainPlotUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
the panel for Density plots;
}
\description{
main bar plots UI.
}
\note{
\code{getBarMainPlotUI}
}
\examples{
    x <- getBarMainPlotUI("bar")

}
