% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxmain.R
\name{getBoxMainPlotUI}
\alias{getBoxMainPlotUI}
\title{getBoxMainPlotUI}
\usage{
getBoxMainPlotUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
the panel for Density plots;
}
\description{
main Box plots UI.
}
\note{
\code{getBoxMainPlotUI}
}
\examples{
    x <- getBoxMainPlotUI("box")

}
