% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getBSTableUI}
\alias{getBSTableUI}
\title{getBSTableModal
prepares a Modal to put a table}
\usage{
getBSTableUI(name = NULL, label = NULL, trigger = NULL, size = "large",
  modal = NULL)
}
\arguments{
\item{name, }{name}

\item{label, }{label}

\item{trigger, }{trigger button name}

\item{size, }{size of the modal}

\item{modal, }{table location (in the modal or not)}
}
\value{
the modal
}
\description{
getBSTableModal
prepares a Modal to put a table
}
\examples{
    x<- getBSTableModal()

}
