% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getConditionSelector}
\alias{getConditionSelector}
\title{getConditionSelector}
\usage{
getConditionSelector(num = 0, choices = NULL, selected = NULL)
}
\arguments{
\item{num, }{panel that is going to be shown}

\item{choices, }{sample list}

\item{selected, }{selected smaple list}
}
\description{
Selects user input conditions to run in DESeq.
}
\examples{
    x <- getConditionSelector()

}
