% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getConditionSelectorFromMeta}
\alias{getConditionSelectorFromMeta}
\title{getConditionSelectorFromMeta}
\usage{
getConditionSelectorFromMeta(metadata = NULL, input = NULL, index = 1,
  num = 0, choices = NULL, selected = NULL)
}
\arguments{
\item{metadata, }{meta data table}

\item{input, }{input}

\item{index, }{index}

\item{num, }{num}

\item{choices, }{choices}

\item{selected, }{selected}
}
\description{
Selects user input conditions to run in DESeq from metadata
}
\examples{
    x <- getConditionSelectorFromMeta()

}
