% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getDataForTables}
\alias{getDataForTables}
\title{getDataForTables
get data to fill up tables tab}
\usage{
getDataForTables(input = NULL, init_data = NULL, filt_data = NULL,
  selected = NULL, getMostVaried = NULL, mergedComp = NULL,
  explainedData = NULL)
}
\arguments{
\item{input, }{input parameters}

\item{init_data, }{initial dataset}

\item{filt_data, }{filt_data}

\item{selected, }{selected genes}

\item{getMostVaried, }{most varied genes}

\item{mergedComp, }{merged comparison set}

\item{explainedData, }{pca gene set}
}
\value{
data
}
\description{
getDataForTables
get data to fill up tables tab
}
\examples{
    x <- getDataForTables()

}

