% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getDomains}
\alias{getDomains}
\title{get domains for the main plots}
\usage{
getDomains(filt_data = NULL)
}
\arguments{
\item{filt_data, }{data to get the domains}
}
\value{
domains
}
\description{
get domains for the main plots
}
\examples{
    x<-getDomains()
}

