% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getDown}
\alias{getDown}
\title{getDown
get down regulated data}
\usage{
getDown(filt_data = NULL)
}
\arguments{
\item{filt_data, }{filt_data}
}
\value{
data
}
\description{
getDown
get down regulated data
}
\examples{
    x <- getDown()

}
