% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{getEnrichDO}
\alias{getEnrichDO}
\title{getEnrichDO}
\usage{
getEnrichDO(genelist, pvalueCutoff = 0.01)
}
\arguments{
\item{genelist, }{gene list}

\item{pvalueCutoff, }{the p value cutoff}
}
\value{
enriched DO
}
\description{
getEnrichDO
}
\note{
\code{getEnrichDO}
}
\examples{
    genelist<-getGeneList(c('OCLN', 'ABCC2'))
    x <- getEnrichDO(genelist, NULL)

}

