% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{getEnrichKEGG}
\alias{getEnrichKEGG}
\title{getEnrichKEGG}
\usage{
getEnrichKEGG(genelist, pvalueCutoff = 0.01)
}
\arguments{
\item{genelist, }{gene list}

\item{pvalueCutoff, }{the p value cutoff}
}
\value{
Enriched KEGG
}
\description{
getEnrichKEGG
}
\note{
\code{getEnrichKEGG}
}
\examples{
    genelist<-getGeneList(c('OCLN', 'ABCC2'))
    x <- getEnrichKEGG(genelist,NULL)
}

