% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{getEntrezIds}
\alias{getEntrezIds}
\title{getGeneList}
\usage{
getEntrezIds(genes = NULL, org = "org.Hs.eg.db")
}
\arguments{
\item{genes, }{gene list with fold changes}

\item{org, }{orgranism for gene symbol entrez ID conversion}
}
\value{
ENTREZ ID list
}
\description{
Gathers the gene list to use for GOTerm analysis.
}
\note{
\code{GOTerm}

\code{getEntrezIds}
symobol to ENTREZ ID conversion
}
\examples{
    x <- getEntrezIds()

}
