% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{getEntrezTable}
\alias{getEntrezTable}
\title{getEntrezTable}
\usage{
getEntrezTable(genes = NULL, dat = NULL, org = "org.Hs.eg.db")
}
\arguments{
\item{genes, }{gene list}

\item{dat, }{data matrix}

\item{org, }{orgranism for gene symbol entrez ID conversion}
}
\value{
table with the entrez IDs in the rownames
}
\description{
Gathers the entrezIds of the genes in given list and their data
}
\note{
\code{GOTerm}

\code{getEntrezTable}
symobol to ENTREZ ID conversion
}
\examples{
    x <- getEntrezTable()

}
