% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getGOLeftMenu}
\alias{getGOLeftMenu}
\title{getGOLeftMenu GO Left menu}
\usage{
getGOLeftMenu()
}
\value{
returns the left menu according to the selected tab;
}
\description{
getGOLeftMenu GO Left menu
}
\note{
\code{getGOLeftMenu}
}
\examples{
    x <- getGOLeftMenu()
}

