% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gopanel.R
\name{getGoPanel}
\alias{getGoPanel}
\title{getGoPanel, go term analysis panel}
\usage{
getGoPanel(flag = FALSE)
}
\arguments{
\item{flag, }{flag to show the element in the ui}
}
\value{
the panel for go term analysis;
}
\description{
getGoPanel, go term analysis panel
}
\note{
\code{getGoPanel}
}
\examples{
 
    x <- getGoPanel()

}

