% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gopanel.R
\name{getGOPlots}
\alias{getGOPlots}
\title{getGOPlots, go term analysis panel}
\usage{
getGOPlots(dataset, inputGO = NULL, genelist = list(1))
}
\arguments{
\item{dataset, }{the dataset used}

\item{inputGO, }{input Gene Ontology}

\item{genelist, }{list of genes}
}
\value{
the panel for go plots;
}
\description{
getGOPlots, go term analysis panel
}
\note{
\code{getGOPlots}
}
\examples{
    x<- getGOPlots(mtcars)

}

