% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_props.R
\name{getHoverPlots}
\alias{getHoverPlots}
\title{getHoverPlots}
\usage{
getHoverPlots(bardata = NULL, genename = NULL)
}
\arguments{
\item{bardata, }{barplot data}

\item{genename, }{gene name in the barplots}
}
\description{
Prepares the plots going to be shown when a gene hovered
in the main plots
}
\examples{
    getHoverPlots()

}
