% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getInitialMenu}
\alias{getInitialMenu}
\title{getInitialMenu Initial menu}
\usage{
getInitialMenu(input = NULL, output = NULL, session = NULL)
}
\arguments{
\item{input, }{input from user}

\item{output, }{output to user}

\item{session, }{session info}
}
\value{
returns the initial menu
}
\description{
getInitialMenu Initial menu
}
\note{
\code{getInitialMenu}
}
\examples{
    x <- getInitialMenu()
}

