% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCSection.R
\name{getIntHeatmapVis}
\alias{getIntHeatmapVis}
\title{getIntHeatmapVis}
\usage{
getIntHeatmapVis(randstr = NULL)
}
\arguments{
\item{randstr, }{randstr}
}
\value{
the panel interactive heatmap
}
\description{
Gathers the conditional panel for interactive heatmap
}
\note{
\code{getIntHeatmapVis}
}
\examples{
    x <- getIntHeatmapVis()

}
