% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark.R
\name{getJsonObj}
\alias{getJsonObj}
\title{getJsonObj}
\usage{
getJsonObj(session = NULL, input = NULL, access_token = NULL)
}
\arguments{
\item{session, }{session}

\item{input, }{input}

\item{access_token, }{access_token}
}
\description{
getJsonVars
}
\examples{
    x <- getJsonObj()

}
