% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSection.R
\name{getLeftMenu}
\alias{getLeftMenu}
\title{Left menu for additional plots}
\usage{
getLeftMenu()
}
\value{
returns the left menu according to the selected tab;
}
\description{
Left menu for additional plots
}
\note{
\code{getLeftMenu}
}
\examples{
    x <- getLeftMenu()

}

