% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{getLegendColors}
\alias{getLegendColors}
\title{getLegendColors}
\usage{
getLegendColors(Legend = c("up", "down", "NS"))
}
\arguments{
\item{Legend, }{unique Legends}
}
\value{
mainPlotControls
}
\description{
Generates colors according to the data
}
\note{
\code{getLegendColors}
}
\examples{
    x <- getLegendColors(c("up", "down", "GS", "NS"))
}
