% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{getLegendRadio}
\alias{getLegendRadio}
\title{getLegendRadio}
\usage{
getLegendRadio(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
radio control
}
\description{
Radio buttons for the types in the legend
}
\note{
\code{getLegendRadio}
}
\examples{
   
    x <- getLegendRadio("deprog")

}
