% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getLegendSelect}
\alias{getLegendSelect}
\title{getLegendSelect}
\usage{
getLegendSelect(id = "pca")
}
\arguments{
\item{id, }{namespace id}
}
\description{
select legend
}
\note{
\code{getLegendSelect}
}
\examples{
    x <- getLegendSelect("pca")
}
