% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{getLevelOrder}
\alias{getLevelOrder}
\title{getLevelOrder}
\usage{
getLevelOrder(Level = c("up", "down", "NS"))
}
\arguments{
\item{Level, }{factor levels shown in the legend}
}
\value{
order
}
\description{
Generates the order of the overlapping points
}
\note{
\code{getLevelOrder}
}
\examples{
    x <- getLevelOrder(c("up", "down", "GS", "NS"))
}
