% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getLoadingMsg}
\alias{getLoadingMsg}
\title{getLoadingMsg and gif}
\usage{
getLoadingMsg()
}
\value{
loading msg
}
\description{
getLoadingMsg and gif
}
\note{
\code{getLoadingMsg}
}
\examples{
    x <- getLoadingMsg()
}

