% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainpanel.R
\name{getMainPanel}
\alias{getMainPanel}
\title{getMainPanel, main panel for volcano, scatter and maplot.  
Barplot and box plots are in this page too}
\usage{
getMainPanel(randstr = NULL)
}
\arguments{
\item{randstr, }{random string for the plot containers}
}
\value{
the panel for main plots;
}
\description{
getMainPanel, main panel for volcano, scatter and maplot.  
Barplot and box plots are in this page too
}
\note{
\code{getMainPanel}
}
\examples{
    x <- getMainPanel()

}

