% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainpanel.R
\name{getMainPanelPlots}
\alias{getMainPanelPlots}
\title{getMainPanelPlots}
\usage{
getMainPanelPlots(filt_data = NULL, cols = NULL, conds = NULL,
  input = NULL, compselect = NULL)
}
\arguments{
\item{filt_data, }{filtered data}

\item{cols, }{selected columns}

\item{conds, }{seleced conditions}

\item{input, }{input from ui}

\item{compselect, }{selected comparison number}
}
\value{
panel
}
\description{
Gathers the the plots to be used within the main panel.
}
\examples{
    x <- getMainPanelPlots()

}
