% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{getMainPlotsLeftMenuNew}
\alias{getMainPlotsLeftMenuNew}
\title{getMainPlotsLeftMenuNew}
\usage{
getMainPlotsLeftMenuNew(id)
}
\value{
returns the left menu according to the selected tab;
}
\description{
Generates the Main PLots Left menu to be displayed within the DEBrowser.
}
\note{
\code{getMainPlotsLeftMenuNew}
}
\examples{
    x <- getMainPlotsLeftMenuNew()
}
