% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{getMainPlotUI}
\alias{getMainPlotUI}
\title{getMainPlotUI}
\usage{
getMainPlotUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
the panel for main plots;
}
\description{
main plot for volcano, scatter and maplot.
}
\note{
\code{getMainPlotUI}
}
\examples{
    x <- getMainPlotUI("main")

}
