% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getMergedComparison}
\alias{getMergedComparison}
\title{getMergedComparison}
\usage{
getMergedComparison(dc = NULL, nc = NULL)
}
\arguments{
\item{dc, }{data container}

\item{nc, }{the number of comparisons}
}
\value{
data
}
\description{
getMergedComparison
}
\examples{
    x <- getMergedComparison()

}

