% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getMostVariedList}
\alias{getMostVariedList}
\title{getMostVariedList}
\usage{
getMostVariedList(datavar = NULL, cols = NULL, topn = 500,
  mincount = 10)
}
\arguments{
\item{datavar, }{loaded dataset}

\item{cols, }{selected columns}

\item{topn, }{most varied records}

\item{mincount, }{total min read count for selected samples}
}
\value{
data
}
\description{
Calculates the most varied genes to be used for specific plots
within the DEBrowser.
}
\examples{
    x <- getMostVariedList()

}
