% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{getNormalizedMatrix}
\alias{getNormalizedMatrix}
\title{getNormalized matrix
it requires edgeR package}
\usage{
getNormalizedMatrix(M, method = "TMM")
}
\arguments{
\item{M, }{numeric matrix}

\item{method, }{normalization method for edgeR. default is TMM}
}
\value{
normalized matrix
}
\description{
getNormalized matrix
it requires edgeR package
}
\note{
\code{getGoPanel}
}
\examples{
    x <- getNormalizedMatrix(mtcars)

}

