% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gopanel.R
\name{getOrganism}
\alias{getOrganism}
\title{getOrganism}
\usage{
getOrganism(org)
}
\arguments{
\item{org, }{organism}
}
\value{
organism name for keg
}
\description{
getOrganism
}
\note{
\code{getOrganism}
}
\examples{
    x <- getOrganism()

}
