% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gopanel.R
\name{getOrganismBox}
\alias{getOrganismBox}
\title{getOrganismBox}
\usage{
getOrganismBox()
}
\value{
selectInput
}
\description{
Get the organism Box.
}
\note{
\code{getOrganismBox}

\code{getOrganismBox}
makes the organism box
}
\examples{
    x <- getOrganismBox()

}
