% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getPCAcontolUpdatesJS}
\alias{getPCAcontolUpdatesJS}
\title{getPCAcontolUpdatesJS
in the prep menu we have two PCA plots to show how batch effect correction worked. 
One set of PCA input controls updates two PCA plots with this JS.}
\usage{
getPCAcontolUpdatesJS()
}
\value{
the JS for tab updates
}
\description{
getPCAcontolUpdatesJS
in the prep menu we have two PCA plots to show how batch effect correction worked. 
One set of PCA input controls updates two PCA plots with this JS.
}
\examples{
    x<- getTabUpdateJS()

}
