% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getPCAexplained}
\alias{getPCAexplained}
\title{getPCAexplained}
\usage{
getPCAexplained(datasetInput = NULL, cols = NULL, input = NULL)
}
\arguments{
\item{datasetInput, }{selected data}

\item{cols, }{columns}

\item{input, }{from usern)}
}
\value{
explained plot
}
\description{
Creates a more detailed plot using the PCA results from
the selected dataset.
}
\examples{
    x <- getPCAexplained()

}

