% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getPlotArea}
\alias{getPlotArea}
\title{getPlotArea}
\usage{
getPlotArea(input = NULL, session = NULL)
}
\arguments{
\item{input, }{input variables}

\item{session, }{session}
}
\value{
heatmapply/heatmap.2 plot area
}
\description{
returns plot area either for heatmaply or heatmap.2
}
\examples{
    x <- getPlotArea()

}
