% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{getQAText}
\alias{getQAText}
\title{getQAText
Some questions and answers}
\usage{
getQAText()
}
\value{
help text for QA
}
\description{
getQAText
Some questions and answers
}
\examples{
    x<- getQAText()

}
