% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getQCLeftMenu}
\alias{getQCLeftMenu}
\title{getQCLeftMenu}
\usage{
getQCLeftMenu()
}
\value{
QC left menu
}
\description{
getQCLeftMenu
}
\note{
\code{getQCLeftMenu}
}
\examples{
    x <- getQCLeftMenu()
}

