% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCSection.R
\name{getQCPanel}
\alias{getQCPanel}
\title{getQCPanel}
\usage{
getQCPanel(randstr = NULL)
}
\arguments{
\item{randstr, }{randstr}
}
\value{
the panel for QC plots;
}
\description{
Gathers the conditional panel for QC plots
}
\note{
\code{getQCSection}
}
\examples{
    x <- getQCPanel()

}

