% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCSection.R
\name{getQCPlots}
\alias{getQCPlots}
\title{getQCPlots, for quality checks}
\usage{
getQCPlots(dataset = NULL, input = NULL, metadata = NULL,
  clustering_method = "complete", distance_method = "cor", cex = 2)
}
\arguments{
\item{dataset, }{the dataset to use}

\item{input, }{user input}

\item{metadata, }{coupled samples and conditions}

\item{clustering_method, }{clustering method used}

\item{distance_method, }{distance method used}

\item{cex, }{font size}
}
\value{
the panel for QC plots;
}
\description{
getQCPlots, for quality checks
}
\note{
\code{getQCPlots}
}
\examples{
    x <- getQCPlots()

}

