% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getSampleDetails}
\alias{getSampleDetails}
\title{getSampleDetails}
\usage{
getSampleDetails(output, summary, details, data)
}
\arguments{
\item{output, }{output}

\item{summary, }{summary output name}

\item{details, }{details ouput name}

\item{data, }{data}
}
\value{
panel
}
\description{
get sample details
}
\examples{
    x <- getSampleDetails(output, data)

}
