% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNames.R
\name{getSampleNames}
\alias{getSampleNames}
\title{Prepares initial samples to fill condition boxes.  
it reads the sample names from the data and splits into two.}
\usage{
getSampleNames(dat, part)
}
\arguments{
\item{dat, }{data that have the sample names in the header}

\item{part, }{c(1,2). 1=first half and 2= second half}
}
\value{
sample names.
}
\description{
Prepares initial samples to fill condition boxes.  
it reads the sample names from the data and splits into two.
}
\examples{
    x<-getSampleNames(mtcars, 1)
    x<-getSampleNames(mtcars, 2)
}

