% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getSamples}
\alias{getSamples}
\title{getSamples}
\usage{
getSamples(cnames = NULL, index = 2)
}
\arguments{
\item{cnames, }{names of the  samples}

\item{index, }{starting column in a tab separated file}
}
\value{
choices
}
\description{
Gathers the sample names to be used within DEBrowser.
}
\examples{
    x <- getSamples()

}
