% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getSearchData}
\alias{getSearchData}
\title{getSearchData}
\usage{
getSearchData(dat = NULL, input = NULL)
}
\arguments{
\item{dat, }{table data}

\item{input, }{input params}
}
\value{
data
}
\description{
search the geneset in the tables and return it
}
\examples{
    x <- getSearchData()

}
