% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getSelectedDatasetInput}
\alias{getSelectedDatasetInput}
\title{getSelectedDatasetInput}
\usage{
getSelectedDatasetInput(rdata = NULL, getSelected = NULL,
  getMostVaried = NULL, getGeneSet = NULL, getSelectedDatasetInput = NULL,
  input = NULL)
}
\arguments{
\item{rdata, }{filtered dataset}

\item{getSelected, }{selected data}

\item{getMostVaried, }{most varied data}

\item{getGeneSet, }{given gene set}

\item{getSelectedDatasetInput, }{merged comparison data}

\item{input, }{input parameters}
}
\value{
data
}
\description{
Gathers the user selected dataset output to be displayed.
}
\examples{
    x <- getSelectedDatasetInput()

}

