% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{getSelectInputBox}
\alias{getSelectInputBox}
\title{getSelectInputBox}
\usage{
getSelectInputBox(id = NULL, name = NULL, num = 0, choices = NULL,
  selected = NULL, cw = 2)
}
\arguments{
\item{id, }{input id}

\item{name, }{label of the box}

\item{num, }{panel that is going to be shown}

\item{choices, }{sample list}

\item{selected, }{selected smaple list}

\item{cw, }{column width}
}
\description{
Selects user input conditions to run in DESeq.
}
\examples{
    x <- getSelectInputBox()

}
