% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCSection.R
\name{getShapeColor}
\alias{getShapeColor}
\title{getShapeColor}
\usage{
getShapeColor(input = NULL)
}
\arguments{
\item{input, }{input values}
}
\value{
Color and shape from selection boxes or defaults
}
\description{
Generates the fill and shape selection boxes for PCA plots.
metadata file has to be loaded in this case
}
\examples{
    x <- getShapeColor()
}
