% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getStartPlotsMsg}
\alias{getStartPlotsMsg}
\title{getStartPlotsMsg}
\usage{
getStartPlotsMsg()
}
\value{
return start plot msg
}
\description{
getStartPlotsMsg
}
\note{
\code{getStartPlotsMsg}
}
\examples{
    x <- getStartPlotsMsg()
}

