% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainpanel.R
\name{getStartUp}
\alias{getStartUp}
\title{getStartUp, startup screen}
\usage{
getStartUp()
}
\value{
startup screen
}
\description{
getStartUp, startup screen
}
\note{
\code{getStartUp}
}
\examples{
 
    x <- getStartUp()

}

