% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getTableDetails}
\alias{getTableDetails}
\title{getTableDetails}
\usage{
getTableDetails(output = NULL, session = NULL, tablename = NULL,
  data = NULL, modal = NULL)
}
\arguments{
\item{output, }{output}

\item{session, }{session}

\item{tablename, }{table name}

\item{data, }{matrix data}

\item{modal, }{if it is true, the matrix is going to be in a modal}
}
\value{
panel
}
\description{
get table details
To be able to put a table into two lines are necessary;
into the server part;
getTableDetails(output, session, "dataname", data, modal=TRUE)
into the ui part;
uiOutput(ns("dataname"))
}
\examples{
    x <- getTableDetails()

}
