% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getTableModal}
\alias{getTableModal}
\title{getTableModal
prepares table modal for KEGG}
\usage{
getTableModal()
}
\value{
the info button
}
\description{
getTableModal
prepares table modal for KEGG
}
\examples{
    x<- getTableModal()

}
