% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getTableStyle}
\alias{getTableStyle}
\title{getTableStyle}
\usage{
getTableStyle(dat = NULL, input = NULL, padj = c("padj"),
  foldChange = c("foldChange"), DEsection = TRUE)
}
\arguments{
\item{dat, }{dataset}

\item{input, }{input params}

\item{padj, }{the name of the padj value column in the dataset}

\item{foldChange, }{the name of the foldChange column in the dataset}

\item{DEsection, }{if it is in DESection or not}
}
\description{
User defined selection that selects the style of table to display
within the DEBrowser.
}
\note{
\code{getTableStyle}
}
\examples{
    x <- getTableStyle()
}
