% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getTextOnOff}
\alias{getTextOnOff}
\title{getTextOnOff}
\usage{
getTextOnOff(id = "pca")
}
\arguments{
\item{id, }{namespace id}
}
\description{
text on PCA plot on and off
}
\note{
\code{getTextOnOff}
}
\examples{
    x <- getTextOnOff("pca")
}
