% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getToolTipPCA}
\alias{getToolTipPCA}
\title{getToolTipPCA}
\usage{
getToolTipPCA(dat = NULL)
}
\arguments{
\item{dat, }{data}
}
\value{
tooltip text
}
\description{
Prepares tooltiptext for PCA plot
}
\examples{
x <- getToolTipPCA()

}
