% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getUpDown}
\alias{getUpDown}
\title{getUpDown
get up+down regulated data}
\usage{
getUpDown(filt_data = NULL)
}
\arguments{
\item{filt_data, }{filt_data}
}
\value{
data
}
\description{
getUpDown
get up+down regulated data
}
\examples{
    x <- getUpDown()

}
