% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getVariationData}
\alias{getVariationData}
\title{getVariationData}
\usage{
getVariationData(inputdata = NULL, cols = NULL, conds = NULL,
  key = NULL)
}
\arguments{
\item{inputdata, }{dataset}

\item{cols, }{columns}

\item{conds, }{conditions}

\item{key, }{gene or region name}
}
\value{
plotdata
}
\description{
Adds an id to the data frame being used.
}
\examples{
    x <- getVariationData()

}
