% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapControlsUI}
\alias{heatmapControlsUI}
\title{heatmapControlsUI}
\usage{
heatmapControlsUI(id)
}
\arguments{
\item{id, }{module ID}
}
\value{
HeatmapControls
}
\description{
Generates the left menu to be used for heatmap plots
}
\note{
\code{heatmapControlsUI}
}
\examples{
    x <- heatmapControlsUI("heatmap")
}
