% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{hideObj}
\alias{hideObj}
\title{hideObj}
\usage{
hideObj(btns = NULL)
}
\arguments{
\item{btns, }{hide group of objects with shinyjs}
}
\description{
Hides a shiny object.
}
\examples{
    x <- hideObj()
}
