% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogramControlsUI}
\alias{histogramControlsUI}
\title{histogramControlsUI}
\usage{
histogramControlsUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
returns the left menu
}
\description{
Generates the controls in the left menu for a histogram
}
\note{
\code{histogramControlsUI}
}
\examples{
    x <- histogramControlsUI("histogram")
}
