% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installpack.R
\name{installpack}
\alias{installpack}
\title{installpack}
\usage{
installpack(package_name = NULL, github = FALSE)
}
\arguments{
\item{package_name, }{package name to be installed}

\item{github, }{if github = true}
}
\description{
install packages if they don't exist
display.
}
\note{
\code{installpack}
}
\examples{
    x <- installpack()

}
