% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQR.R
\name{IQRPlotControlsUI}
\alias{IQRPlotControlsUI}
\title{IQRPlotControlsUI}
\usage{
IQRPlotControlsUI(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
returns the left menu
}
\description{
Generates the controls in the left menu for an IQR plot#'
}
\note{
\code{IQRPlotControlsUI}
}
\examples{
    x <- IQRPlotControlsUI("IQR")
}
