% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowcountfilter.R
\name{lcfMetRadio}
\alias{lcfMetRadio}
\title{lcfMetRadio}
\usage{
lcfMetRadio(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
radio control
}
\description{
Radio buttons for low count removal methods
}
\note{
\code{lcfMetRadio}
}
\examples{
   
    x <- lcfMetRadio("lcf")

}
