% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_brush.R
\name{link_brush}
\alias{link_brush}
\title{link_brush}
\usage{
link_brush()
}
\value{
A list with components:
    \item{input}{A function that takes a visualisation as an argument and
     adds an input brush to that plot}
    \item{selected}{A reactive providing a logical vector that describes
    which points are under the brush}
}
\description{
Modified linked brush object.
A link brush function modified to be able to create non-reactive
linked brush object for ggvis plots
}
\note{
\code{link_brush} is very new and is likely to change
    substantially
}
\examples{
    lb <- link_brush()

}
