% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delinked_brush.R
\name{linked_brush_ma}
\alias{linked_brush_ma}
\title{Modified linked brush object.}
\usage{
linked_brush_ma()
}
\value{
A list with components:
    \item{input}{A function that takes a visualisation as an argument and
    adds an input brush to that plot}
    \item{selected}{A reactive providing a logical vector that describes
    which points are under the brush}
}
\description{
A linked brush function modified to be able to create non-reactive
linked brush object for ggvis plots
}
\note{
\code{delinked_brush} is very new and is likely to change
    substantially
}
\examples{
    lb <- linked_brush_ma()
}

