% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data.}
\usage{
load_data(input = NULL, session = NULL)
}
\arguments{
\item{input, }{input values}

\item{session, }{if data is going to be loaded from json}
}
\value{
data
}
\description{
Load data.
}
\examples{
    x<-load_data ()
}

