% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{mainScatterNew}
\alias{mainScatterNew}
\title{mainScatterNew}
\usage{
mainScatterNew(input = NULL, data = NULL, source = NULL, x = NULL,
  y = NULL)
}
\arguments{
\item{data, }{dataframe that has log2FoldChange and log10padj values}

\item{source, }{for event triggering to select genes}

\item{x, }{the name of the x coordinate}

\item{y, }{the name of the y coordinate}
}
\value{
scatter, volcano or MA plot
}
\description{
Creates the main scatter, volcano or MA plot to be displayed within the main
panel.
}
\examples{
    
    x <- mainScatterNew(input, data, source, x, y)

}
