% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPlot.R
\name{MAPlot}
\alias{MAPlot}
\title{Prepares volcano plot}
\usage{
MAPlot(dat = NULL, lb = NULL, data_tooltip = NULL)
}
\arguments{
\item{dat, }{dataframe that has log2FoldChange and log10padj values}

\item{lb, }{the linked brush}

\item{data_tooltip, }{toolstip specific to this plot}
}
\value{
MA plot
}
\description{
Prepares volcano plot
}
\examples{
    x <- MAPlot()

}

