% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPlot.R
\name{MAZoom}
\alias{MAZoom}
\title{MA Zoom}
\usage{
MAZoom(dat = NULL, data_tooltip = NULL)
}
\arguments{
\item{dat, }{dataframe that has log2FoldChange and log10padj values}

\item{data_tooltip, }{toolstip specific to this plot}
}
\value{
zoomed MA plot
}
\description{
MA Zoom
}
\examples{
    x <- MAZoom()

}

