% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batcheffect.R
\name{normalizationMethods}
\alias{normalizationMethods}
\title{normalizationMethods}
\usage{
normalizationMethods(id)
}
\arguments{
\item{id, }{namespace id}
}
\value{
radio control
}
\description{
Select box to select normalization method prior to batch effect correction
}
\note{
\code{normalizationMethods}
}
\examples{
   
    x <- normalizationMethods("batch")

}
