% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{plotData}
\alias{plotData}
\title{plotData}
\usage{
plotData(pdata, input)
}
\arguments{
\item{data, }{data}
}
\value{
prepdata
}
\description{
prepare plot data for mainplots
}
\note{
\code{plotData}
}
\examples{
    x <- plotData(data)
}
