% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSize.R
\name{plotMarginsUI}
\alias{plotMarginsUI}
\title{plotMarginsUI}
\usage{
plotMarginsUI(id, t = 20, b = 100, l = 100, r = 20)
}
\arguments{
\item{id, }{id}

\item{t, }{top margin}

\item{b, }{bottom margin}

\item{l, }{left margin}

\item{r, }{right margin}
}
\value{
size and margins controls
}
\description{
Margins module for plotly plots
}
\note{
\code{plotMarginsUI}
}
\examples{
    x <- plotMarginsUI("heatmap")
}
