% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCSection.R
\name{prepAddQCPlots}
\alias{prepAddQCPlots}
\title{prepAddQCPlots}
\usage{
prepAddQCPlots(data = NULL, input = NULL)
}
\arguments{
\item{data, }{barplot data}

\item{input, }{user input params}
}
\description{
prepares IQR and density plots
}
\examples{
    prepAddQCPlots()


}
