% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{prepDataContainerNew}
\alias{prepDataContainerNew}
\title{prepDataContainerNew}
\usage{
prepDataContainerNew(data = NULL, counter = NULL, input = NULL)
}
\arguments{
\item{data, }{loaded dataset}

\item{counter, }{the number of comparisons}

\item{input, }{input parameters}
}
\value{
data
}
\description{
Prepares the data container that stores values used within DESeq.
}
\examples{
    x <- prepDataContainerNew()

}
