% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{prepDEOutput}
\alias{prepDEOutput}
\title{prepDEOutput}
\usage{
prepDEOutput(data = NULL, cols = NULL, conds = NULL, inputconds = NULL,
  i = NULL, input = NULL)
}
\arguments{
\item{data, }{loaded dataset}

\item{cols, }{columns}

\item{conds, }{conds}

\item{inputconds, }{inputconds}

\item{i, }{selected comparison number}

\item{input, }{input}
}
\value{
data
}
\description{
Prepares the output data from DE analysis to be used within
DEBrowser
}
\examples{
    x <- prepDEOutput()

}
