% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{prepDESeqOutput}
\alias{prepDESeqOutput}
\title{prepDESeqOutput}
\usage{
prepDESeqOutput(data = NULL, cols = NULL, conds = NULL,
  inputconds = NULL, i = NULL)
}
\arguments{
\item{data, }{loaded dataset}

\item{cols, }{columns}

\item{conds, }{conds}

\item{inputconds, }{inputconds}

\item{i, }{selected comparison number}
}
\value{
data
}
\description{
Prepares the output data from DESeq to be used within
DEBrowser
}
\examples{
    x <- prepDESeqOutput()

}

