% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{prepGroup}
\alias{prepGroup}
\title{prepGroup}
\usage{
prepGroup(conds = NULL, cols = NULL)
}
\arguments{
\item{conds, }{inputconds}

\item{cols, }{columns}
}
\value{
data
}
\description{
prepare group table
}
\examples{
    x <- prepGroup()

}
