% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{prepHeatData}
\alias{prepHeatData}
\title{prepHeatData}
\usage{
prepHeatData(data)
}
\arguments{
\item{data, }{a matrixthat includes expression values}
}
\value{
heatdata
}
\description{
scales the data
}
\examples{
    x <- prepHeatData(mtcars)

}
