% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{readMetaData}
\alias{readMetaData}
\title{readMetaData}
\usage{
readMetaData(input = NULL)
}
\arguments{
\item{input, }{input values}
}
\description{
read metadata file
}
\note{
\code{readMetaData}
}
\examples{
    x <- readMetaData()
}
