% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookmark.R
\name{removeBookmark}
\alias{removeBookmark}
\title{removeBookmark}
\usage{
removeBookmark(ID = NULL, username = NULL)
}
\arguments{
\item{ID, }{prev url}

\item{username, }{username}
}
\description{
remove saved state
}
\examples{
    x <- removeBookmark()

}
