% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{round_vals}
\alias{round_vals}
\title{round_vals}
\usage{
round_vals(l)
}
\arguments{
\item{l, }{the value}
}
\value{
round value
}
\description{
Plot PCA results.
}
\examples{
    x<-round_vals(5.1323223)
}
