% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{run_pca}
\alias{run_pca}
\title{run_pca}
\usage{
run_pca(x = NULL, retx = TRUE, center = TRUE, scale = TRUE)
}
\arguments{
\item{x, }{dataframe with experiment data}

\item{retx, }{specifies if the data should be returned}

\item{center, }{center the PCA (Boolean)}

\item{scale, }{scale the PCA (Boolean)}
}
\value{
pca list
}
\description{
Runs PCA on the selected dataset.
}
\examples{
    load(system.file("extdata", "demo", "demodata.Rda", 
        package="debrowser"))
    pca_data<-run_pca(getNormalizedMatrix(
        demodata[rowSums(demodata[,2:7])>10,2:7]))

}
