% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{runBayseq}
\alias{runBayseq}
\title{runBayseq}
\usage{
runBayseq(data = NULL, columns = NULL, conds = NULL, rowsum.filter = 10)
}
\arguments{
\item{data, }{A matrix that includes all the expression raw counts,
rownames has to be the gene, isoform or region names/IDs}

\item{columns, }{is a vector that includes the columns that are going
to be analyzed. These columns has to match with the given data.}

\item{conds, }{experimental conditions. The order has to match
with the column order}

\item{rowsum.filter, }{regions/genes/isoforms with total count 
(across all samples) below this value will be filtered out}
}
\value{
BaySeq results
}
\description{
Run Bayseq algorithm on the selected conditions.  Output is
to be used for the interactive display.
}
\examples{
    x <- runBayseq()

}
