% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2.R
\name{runDESeq}
\alias{runDESeq}
\title{runDESeq}
\usage{
runDESeq(data, columns, conds, fitType = c("parametric", "local", "mean"),
  non_expressed_cutoff = 10)
}
\arguments{
\item{data, }{A matrix that includes all the expression raw counts,
rownames has to be the gene, isoform or region names/IDs}

\item{columns, }{is a vector that includes the columns that are going
to be analyzed. These columns has to match with the given data.}

\item{conds, }{experimental conditions. The order has to match
with the column order}

\item{fitType, }{DESeq2 fitType, it can be 'parametric', 
'local', 'mean'.}

\item{non_expressed_cutoff, }{to remove unexpressed 
regions/genes/isoforms this cutoff is used}
}
\value{
deseq2 results
}
\description{
Run DESeq2 algorithm on the selected conditions.  Output is
to be used for the interactive display.
}
\examples{
    x <- runDESeq(data<-NULL, columns<-c())

}
