% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{runHeatmap2}
\alias{runHeatmap2}
\title{runHeatmap2}
\usage{
runHeatmap2(input = NULL, session = NULL, data = NULL)
}
\arguments{
\item{input, }{input variables}

\item{session, }{session}

\item{data, }{a matrix that includes expression values}
}
\value{
heatmap.2
}
\description{
Creates a heatmap based on the user selected parameters within shiny
}
\examples{
    x <- runHeatmap2()

}
