% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{scatterZoom}
\alias{scatterZoom}
\title{scatter zoom}
\usage{
scatterZoom(dat = NULL, data_tooltip = NULL)
}
\arguments{
\item{dat, }{dataframe that has log2FoldChange and log10padj values}

\item{data_tooltip, }{toolstip specific to this plot}
}
\value{
zoomed scatter plot
}
\description{
scatter zoom
}
\examples{
    x <- scatterZoom()

}

