% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{selectBatchEffect}
\alias{selectBatchEffect}
\title{selectBatchEffect}
\usage{
selectBatchEffect(input = NULL)
}
\arguments{
\item{input, }{input values}
}
\description{
Batch effect column selection
}
\note{
\code{selectBatchEffect}
}
\examples{
    x <- selectBatchEffect()
}

