% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{selectConditions}
\alias{selectConditions}
\title{selectConditions, multiple condition selection}
\usage{
selectConditions(Dataset = NULL, choicecounter, input = NULL)
}
\arguments{
\item{Dataset, }{used dataset}

\item{choicecounter, }{total number of comparisons}

\item{input, }{input params}
}
\value{
the panel for go plots;
}
\description{
selectConditions, multiple condition selection
}
\note{
\code{selectConditions}
}
\examples{
    x<- selectConditions()

}

