% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{selectGroupInfo}
\alias{selectGroupInfo}
\title{selectGroupInfo}
\usage{
selectGroupInfo(metadata = NULL, input = NULL, selectname = "groupselect",
  label = "Group info")
}
\arguments{
\item{input, }{input values}

\item{selectname, }{name of the select box}

\item{label, }{label of the select box}
}
\description{
Group info column selection. This can be used in batch effect
or coloring the groups in the plots.
}
\note{
\code{selectGroupInfo}
}
\examples{
    x <- selectGroupInfo()
}
