% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R
\name{setFilterParamsNew}
\alias{setFilterParamsNew}
\title{setFilterParamsNew}
\usage{
setFilterParamsNew(session = NULL, input = NULL)
}
\arguments{
\item{session, }{session variable}

\item{input, }{input parameters}
}
\description{
It sets the filter parameters
}
\examples{
    x <- setFilterParamsNew()

}
