% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{showObj}
\alias{showObj}
\title{showObj}
\usage{
showObj(btns = NULL)
}
\arguments{
\item{btns, }{show group of objects with shinyjs}
}
\description{
Displays a shiny object.
}
\examples{
    x <- showObj()
}
