% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startShiny.R
\name{startDEBrowser}
\alias{startDEBrowser}
\title{startDEBrowser to be able to run interactively}
\usage{
startDEBrowser()
}
\value{
the app
}
\description{
startDEBrowser to be able to run interactively
}
\note{
\code{startDEBrowser}
}
\examples{
    startDEBrowser()

}

