% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{togglePanels}
\alias{togglePanels}
\title{togglePanels}
\usage{
togglePanels(num = NULL, nums = NULL, session = NULL)
}
\arguments{
\item{num, }{selected panel}

\item{nums, }{all panels}

\item{session, }{session info}
}
\description{
togglePanels
}
\note{
\code{togglePanels}
}
\examples{
    x <- togglePanels()
}

