% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Prepares volcano plot}
\usage{
volcanoPlot(dat = NULL, lb = NULL, data_tooltip = NULL)
}
\arguments{
\item{dat, }{dataframe that has log2FoldChange and log10padj values}

\item{lb, }{the linked brush}

\item{data_tooltip, }{toolstip specific to this plot}
}
\value{
volcano plot
}
\description{
Prepares volcano plot
}
\examples{
    x <- volcanoPlot()

}

