% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcanoZoom}
\alias{volcanoZoom}
\title{Prepares volcano plot}
\usage{
volcanoZoom(dat = NULL, data_tooltip = NULL)
}
\arguments{
\item{dat, }{dataframe that has log2FoldChange and log10padj values}

\item{data_tooltip, }{toolstip specific to this plot}
}
\value{
zoomed volcano plot
}
\description{
Prepares volcano plot
}
\examples{
    x <- volcanoZoom()

}

