% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{x}
\alias{x}
\title{Global variable x}
\format{An object of class \code{character} of length 1.}
\usage{
x
}
\value{
x
}
\description{
Global variable x for condition 1
}
\examples{
    x <- "Cond1"

}
\keyword{datasets}

