% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{y}
\alias{y}
\title{Global variable y}
\format{An object of class \code{character} of length 1.}
\usage{
y
}
\value{
y
}
\description{
Global variable y for condition 2
}
\examples{
    y <- "Cond2"

}
\keyword{datasets}

