% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{addDataCols}
\alias{addDataCols}
\title{addDataCols}
\usage{
addDataCols(data = NULL, de_res = NULL, cols = NULL, inputconds = NULL,
  i = NULL, input = NULL)
}
\arguments{
\item{data, }{loaded dataset}

\item{de_res, }{de results}

\item{cols, }{columns}

\item{inputconds, }{inputconds}

\item{i, }{selected comparison number}

\item{input, }{input}
}
\value{
data
}
\description{
add aditional data columns to de results
}
\examples{
    x <- addDataCols()

}
