% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprogs.R, R/prepdata.R
\name{applyFilters}
\alias{applyFilters}
\alias{applyFilters}
\title{applyFilters}
\usage{
applyFilters(filt_data = NULL, cols = NULL, conds = NULL, input = NULL)

applyFilters(filt_data = NULL, cols = NULL, conds = NULL, input = NULL)
}
\arguments{
\item{filt_data, }{loaded dataset}

\item{cols, }{selected samples}

\item{conds, }{seleced conditions}

\item{input, }{input parameters}

\item{data, }{loaded dataset}

\item{input, }{input parameters}
}
\value{
data

data
}
\description{
Apply filters based on foldChange cutoff and padj value.
This function adds a "Legend" column with "Up", "Down" or
"NS" values for visualization.

Applies filters based on user selected parameters to be
displayed within the DEBrowser.
}
\examples{
    x <- applyFilters()

    x <- applyFilters()

}
