% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{clusterData}
\alias{clusterData}
\title{clusterData}
\usage{
clusterData(dat = NULL)
}
\arguments{
\item{dat, }{the data to cluster}
}
\value{
clustered data
}
\description{
Gathers the Cluster analysis data to be used within the
GO Term plots.
}
\note{
\code{clusterData}
}
\examples{
    mycluster <- clusterData()

}
