% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{deUI}
\alias{deUI}
\title{deUI}
\usage{
deUI()
}
\value{
the panel for main plots;
}
\description{
Creates a shinyUI to be able to run DEBrowser interactively.
}
\note{
\code{deUI}
}
\examples{
    x<-deUI()

}
