% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getAfterLoadMsg}
\alias{getAfterLoadMsg}
\title{getAfterLoadMsg}
\usage{
getAfterLoadMsg()
}
\value{
return After Load Msg
}
\description{
Generates and displays the message to be shown after loading data
within the DEBrowser.
}
\note{
\code{getAfterLoadMsg}
}
\examples{
    x <- getAfterLoadMsg()
}
