% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getBSTableUI}
\alias{getBSTableUI}
\title{getBSTableModal
prepares a Modal to put a table}
\usage{
getBSTableUI(name, label, trigger, size = "large", modal = NULL)
}
\value{
the modal
}
\description{
getBSTableModal
prepares a Modal to put a table
}
\examples{
    x<- getBSTableModal()

}
