% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getColorShapeSelection}
\alias{getColorShapeSelection}
\title{getColorShapeSelection}
\usage{
getColorShapeSelection(input = NULL)
}
\arguments{
\item{input, }{input values}
}
\value{
Color and shape selection boxes
}
\description{
Generates the fill and shape selection boxes for PCA plots.
metadata file has to be loaded in this case
}
\examples{
    x <- getColorShapeSelection()
}
