% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{getCompSelection}
\alias{getCompSelection}
\title{getCompSelection}
\usage{
getCompSelection(name = NULL, count = NULL)
}
\arguments{
\item{name, }{the name of the selectInput}

\item{count, }{comparison count}
}
\description{
Gathers the user selected comparison set to be used within the
DEBrowser.
}
\note{
\code{getCompSelection}
}
\examples{
    x <- getCompSelection(name="comp", count = 2)
}
