% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getCompSelection}
\alias{getCompSelection}
\title{getCompSelection}
\usage{
getCompSelection(count = NULL)
}
\arguments{
\item{count, }{comparison count}
}
\description{
Gathers the user selected comparison set to be used within the
DEBrowser.
}
\note{
\code{getCompSelection}
}
\examples{
    x <- getCompSelection(count = 2)
}
