% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getCutOffSelection}
\alias{getCutOffSelection}
\title{getCutOffSelection}
\usage{
getCutOffSelection(nc = 1)
}
\arguments{
\item{nc, }{total number of comparisons}
}
\value{
returns the left menu according to the selected tab;
}
\description{
Gathers the cut off selection for DE analysis
}
\note{
\code{getCutOffSelection}
}
\examples{
    x <- getCutOffSelection()
}
