% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getDataPrepPanel}
\alias{getDataPrepPanel}
\title{getDataPrepPanel}
\usage{
getDataPrepPanel()
}
\value{
returns the prep panel;
}
\description{
Create and show the Condition selection screen to the user
within the DEBrowser.
}
\note{
\code{getDataPrepPanel}
}
\examples{
    x <- getDataPrepPanel()
}
