% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getDataPrepPanel}
\alias{getDataPrepPanel}
\title{getDataPrepPanel}
\usage{
getDataPrepPanel(flag = FALSE)
}
\arguments{
\item{flag, }{flag to show the element in the ui}
}
\value{
returns the left menu according to the selected tab;
}
\description{
Create and show the Condition selection screen to the user
within the DEBrowser.
}
\note{
\code{getDataPrepPanel}
}
\examples{
    x <- getDataPrepPanel()
}
