% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadSection.R
\name{getDownloadSection}
\alias{getDownloadSection}
\title{getDownloadSection, download section button and dataset selection
    box in the menu}
\usage{
getDownloadSection(flag = FALSE, type = "main")
}
\arguments{
\item{flag, }{to show the download selection}

\item{type, }{main vs. QC section}
}
\value{
the panel for download section in the menu;
}
\description{
getDownloadSection, download section button and dataset selection
    box in the menu
}
\note{
\code{getDownloadSection}
}
\examples{
    x<- getDownloadSection()

}

