% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getDownloadSection}
\alias{getDownloadSection}
\title{getDownloadSection}
\usage{
getDownloadSection(choices = NULL)
}
\arguments{
\item{choices, }{main vs. QC section}
}
\value{
the panel for download section in the menu;
}
\description{
download section button and dataset selection box in the
menu for user to download selected data.
}
\note{
\code{getDownloadSection}
}
\examples{
    x<- getDownloadSection()

}
