% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOterm.R
\name{getEnrichDO}
\alias{getEnrichDO}
\title{getEnrichDO}
\usage{
getEnrichDO(genelist = NULL, pvalueCutoff = 0.01)
}
\arguments{
\item{genelist, }{gene list}

\item{pvalueCutoff, }{the p value cutoff}
}
\value{
enriched DO
}
\description{
Gathers the Enriched DO Term analysis data to be used within the
GO Term plots.
}
\note{
\code{getEnrichDO}
}
\examples{
    x <- getEnrichDO()
}
