% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gopanel.R
\name{getGOPlots}
\alias{getGOPlots}
\title{getGOPlots}
\usage{
getGOPlots(dataset, input = NULL, table = FALSE)
}
\arguments{
\item{dataset, }{the dataset used}

\item{input, }{input params}

\item{table, }{table flag to return dataset to show in a datatable}
}
\value{
the panel for go plots;
}
\description{
Go term analysis panel.  Generates appropriate GO plot
based on user selection.
}
\note{
\code{getGOPlots}
}
\examples{
    x<- getGOPlots(mtcars)
}

