% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getIntHeatmap}
\alias{getIntHeatmap}
\title{getIntHeatmap}
\usage{
getIntHeatmap(data = NULL, input = NULL, inputQCPlot = NULL)
}
\arguments{
\item{data, }{heatData}

\item{input, }{all input params}

\item{inputQCPlot, }{input poarams for QC}
}
\value{
plot
}
\description{
getIntHeatmap
}
\examples{
    getIntHeatmap()

}
