% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getLeftMenu}
\alias{getLeftMenu}
\title{getLeftMenu Left menu for plots}
\usage{
getLeftMenu(flag = TRUE)
}
\arguments{
\item{flag, }{flag to show the element in the ui}
}
\value{
returns the left menu according to the selected tab;
}
\description{
getLeftMenu Left menu for plots
}
\note{
\code{getLeftMenu}
}
\examples{
    x <- getLeftMenu()
}

