% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getLeftMenu}
\alias{getLeftMenu}
\title{getLeftMenu}
\usage{
getLeftMenu(input = NULL)
}
\arguments{
\item{input, }{input values}
}
\value{
returns the left menu according to the selected tab;
}
\description{
Generates the left menu for for plots within the DEBrowser.
}
\note{
\code{getLeftMenu}
}
\examples{
    x <- getLeftMenu()
}
