% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getLoadingMsg}
\alias{getLoadingMsg}
\title{getLoadingMsg}
\usage{
getLoadingMsg(output = NULL)
}
\arguments{
\item{output, }{output message}
}
\value{
loading msg
}
\description{
Creates and displays the loading message/gif to be displayed
within the DEBrowser.
}
\note{
\code{getLoadingMsg}
}
\examples{
    x <- getLoadingMsg()
}
