% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getMainPlotsLeftMenu}
\alias{getMainPlotsLeftMenu}
\title{getMainPlotsLeftMenu}
\usage{
getMainPlotsLeftMenu()
}
\value{
returns the left menu according to the selected tab;
}
\description{
Generates the Main PLots Left menu to be displayed within the DEBrowser.
}
\note{
\code{getMainPlotsLeftMenu}
}
\examples{
    x <- getMainPlotsLeftMenu()
}
