% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getMean}
\alias{getMean}
\title{getMean}
\usage{
getMean(norm_data = NULL, de_res = NULL, inputconds = NULL,
  colnum = NULL)
}
\arguments{
\item{norm_data, }{loaded dataset}

\item{de_res, }{de results}

\item{inputconds, }{input parameters}

\item{colnum, }{colnum}
}
\value{
data
}
\description{
Gathers the mean for selected condition.
}
\examples{
    x <- getMean()

}
