% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{getMostVariedList}
\alias{getMostVariedList}
\title{getMostVariedList}
\usage{
getMostVariedList(datavar = NULL, cols = NULL, input = NULL)
}
\arguments{
\item{datavar, }{loaded dataset}

\item{cols, }{selected columns}

\item{input, }{input}
}
\value{
data
}
\description{
Calculates the most varied genes to be used for specific plots
within the DEBrowser.
}
\examples{
    x <- getMostVariedList()

}
