% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{getPCAexplained}
\alias{getPCAexplained}
\title{getPCAexplained}
\usage{
getPCAexplained(datasetInput = NULL, input = NULL)
}
\arguments{
\item{datasetInput, }{selected data}

\item{input, }{from user}
}
\value{
explained plot
}
\description{
Creates a more detailed plot using the PCA results from
the selected dataset.
}
\examples{
load(system.file("extdata", "demo", "demodata.Rda", 
package="debrowser"))
input<-c()
input$qcplot<-"pca"
input$col_list<-colnames(demodata[,2:7])
x <- getPCAexplained(getNormalizedMatrix(demodata[,2:7]), 
    input)

}
