% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getPCselection}
\alias{getPCselection}
\title{getPCselection}
\usage{
getPCselection(num = 1, xy = "x")
}
\arguments{
\item{num, }{PC selection number}

\item{xy, }{x or y coordinate}
}
\value{
PC selection for PCA analysis
}
\description{
Generates the PC selection number to be used within DEBrowser.
}
\note{
\code{getPCselection}
}
\examples{
    x <- getPCselection()
}
