% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{getProgramTitle}
\alias{getProgramTitle}
\title{getProgramTitle}
\usage{
getProgramTitle(session = NULL)
}
\arguments{
\item{session, }{session var}
}
\value{
program title
}
\description{
Generates the title of the program to be displayed within DEBrowser.
If it is called in a program, the program title will be hidden
}
\note{
\code{getProgramTitle}
}
\examples{
    title<-getProgramTitle()
}
