% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCSection.R
\name{getQCPanel}
\alias{getQCPanel}
\title{getQCPanel, conditional panel for QC plots}
\usage{
getQCPanel(flag = FALSE)
}
\arguments{
\item{flag, }{to show the section}
}
\value{
the panel for QC plots;
}
\description{
getQCPanel, conditional panel for QC plots
}
\note{
\code{getQCSection}
}
\examples{
    x <- getQCPanel()

}

