% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCSection.R
\name{getQCPlotArea}
\alias{getQCPlotArea}
\title{getQCPlotArea}
\usage{
getQCPlotArea(input = NULL, flag = FALSE)
}
\arguments{
\item{input, }{user input}

\item{flag, }{flag to show the element in the ui}
}
\description{
Displays the QC plot output area to the user within the
DEBrowser.
}
\examples{
    x <- getQCPlotArea()

}

