% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleNames.R
\name{getSampleNames}
\alias{getSampleNames}
\title{getSampleNames}
\usage{
getSampleNames(cnames = NULL, part = 1)
}
\arguments{
\item{cnames, }{sample names in the header of a dataset}

\item{part, }{c(1,2). 1=first half and 2= second half}
}
\value{
sample names.
}
\description{
Prepares initial samples to fill condition boxes.  
it reads the sample names from the data and splits into two.
}
\examples{
    x<-getSampleNames()
}
