% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{getSelHeat}
\alias{getSelHeat}
\title{getSelHeat}
\usage{
getSelHeat(init_data = NULL, heatdat = NULL, count = NULL)
}
\arguments{
\item{init_data, }{initial data}

\item{heatdat, }{heatData}

\item{count, }{selected gene count}
}
\value{
plot
}
\description{
heatmap selection functionality
}
\examples{
    x <- getSelHeat()

}

