% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{logSliderJScode}
\alias{logSliderJScode}
\title{logSliderJScode}
\usage{
logSliderJScode(slidername = NULL)
}
\arguments{
\item{slidername, }{id of the slider}
}
\value{
returns the slider values in log10 scale
}
\description{
Generates the log based slider to be used by the user within
DEBrowser.
}
\note{
\code{logSliderJScode}
}
\examples{
    x <- logSliderJScode()
}
