% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPlot.R
\name{MAZoom}
\alias{MAZoom}
\title{MAZoom}
\usage{
MAZoom(dat = NULL, data_tooltip = NULL, domains = NULL, colors = NULL)
}
\arguments{
\item{dat, }{dataframe that has log2FoldChange and log10padj values}

\item{data_tooltip, }{toolstip specific to this plot}

\item{domains, }{the domains to be colored}

\item{colors, }{colors for each domain}
}
\value{
zoomed MA plot
}
\description{
Prepares the zoomed in version of the MA plot to be used within
the main panel.
}
\examples{
    x <- MAZoom()

}
