% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{plotData}
\alias{plotData}
\title{plotData}
\usage{
plotData(pdata = NULL, input = NULL)
}
\arguments{
\item{pdata, }{data}

\item{input, }{input}
}
\value{
prepdata
}
\description{
prepare plot data for mainplots
}
\note{
\code{plotData}
}
\examples{
    x <- plotData()
}
