% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{prepDataForQC}
\alias{prepDataForQC}
\title{prepDataForQC}
\usage{
prepDataForQC(dataset = NULL, input = NULL)
}
\arguments{
\item{dataset, }{loaded dataset}

\item{input, }{input}
}
\value{
data
}
\description{
Prepares selected data for QC plots.
}
\examples{
    x <- prepDataForQC()

}
