% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCSection.R
\name{saveQCPlot}
\alias{saveQCPlot}
\title{saveQCPlot}
\usage{
saveQCPlot(filename = NULL, input = NULL, datasetInput = NULL,
  cols = NULL, conds = NULL, inputQCPlot = NULL)
}
\arguments{
\item{filename, }{filename}

\item{input, }{input params}

\item{datasetInput, }{dataset}

\item{cols, }{selected columns}

\item{conds, }{selected conditions}

\item{inputQCPlot, }{clustering method and distance method}
}
\description{
Saves the current QC plot selection to the users local disk.
}
\note{
\code{saveQCPlot}
}
\examples{
    saveQCPlot()

}
