% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainScatter.R
\name{scatterZoom}
\alias{scatterZoom}
\title{scatter zoom}
\usage{
scatterZoom(dat = NULL, data_tooltip = NULL, x = NULL, y = NULL,
  domains = NULL, colors = NULL)
}
\arguments{
\item{dat, }{dataframe that has log2FoldChange and log10padj values}

\item{data_tooltip, }{toolstip specific to this plot}

\item{x, }{the name of the x coordinate}

\item{y, }{the name of the y coordinate}

\item{domains, }{the domains to be colored}

\item{colors, }{colors for each domain}
}
\value{
zoomed scatter plot
}
\description{
scatter zoom
}
\examples{
    x <- scatterZoom()

}

