% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifuncs.R
\name{selectBatchEffect}
\alias{selectBatchEffect}
\title{selectBatchEffect}
\usage{
selectBatchEffect(input = NULL, selectname = "batchselect",
  label = "Batch effect correction column")
}
\arguments{
\item{input, }{input values}

\item{selectname, }{name of the select box}

\item{label, }{label of the select box}
}
\description{
Batch effect column selection
}
\note{
\code{selectBatchEffect}
}
\examples{
    x <- selectBatchEffect()
}
