% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condSelect.R
\name{selectConditions}
\alias{selectConditions}
\title{selectConditions}
\usage{
selectConditions(Dataset = NULL, choicecounter, input = NULL,
  loadingJSON = NULL)
}
\arguments{
\item{Dataset, }{used dataset}

\item{choicecounter, }{total number of comparisons}

\item{input, }{input params}

\item{loadingJSON, }{loads from json}
}
\value{
the panel for go plots;
}
\description{
Selects user input conditions, multiple if present, to be
used in DESeq.
}
\note{
\code{selectConditions}
}
\examples{
    x<- selectConditions()

}
