% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepdata.R
\name{setFilterParams}
\alias{setFilterParams}
\title{setFilterParams}
\usage{
setFilterParams(session = NULL, input = NULL)
}
\arguments{
\item{session, }{session variable}

\item{input, }{input parameters}
}
\description{
It sets the filter parameters
}
\examples{
    x <- setFilterParams()

}
