% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startShiny.R
\name{startDEBrowser}
\alias{startDEBrowser}
\title{startDEBrowser}
\usage{
startDEBrowser()
}
\value{
the app
}
\description{
Starts the DEBrowser to be able to run interactively.
}
\note{
\code{startDEBrowser}
}
\examples{
    startDEBrowser()

}
