% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{volcanoPlot}
\usage{
volcanoPlot(dat = NULL, lb = NULL, data_tooltip = NULL, domains = NULL,
  colors = NULL)
}
\arguments{
\item{dat, }{dataframe that has log2FoldChange and log10padj values}

\item{lb, }{the linked brush}

\item{data_tooltip, }{toolstip specific to this plot}

\item{domains, }{the domains to be colored}

\item{colors, }{colors for each domain}
}
\value{
volcano plot
}
\description{
Prepares volcano plot to be used within the DEBrowser.
}
\examples{
    x <- volcanoPlot()

}
